requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
python-engineio==3.13.2
python-socketio==4.6.0
redis==3.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.25.10
websocket-client==0.57.0
Werkzeug==1.0.1
